auto-mograph-bot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-mograph-bot"
version = "0.1.0"
description = "Automation pipeline for text-to-video generation"
authors = [{ name = "Auto Mograph Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "pydantic>=2.5",
    "rich>=13.7",
    "httpx>=0.25",
    "playwright>=1.40",
    "appium-python-client>=2.11",
    "PySide6>=6.6",
    "keyring>=24.3",
    "SQLAlchemy>=2.0",
    "PyMySQL>=1.1",
    "alembic>=1.12",
    "requests>=2.31",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.1.5",
]

[tool.setuptools.packages.find]
where = ["src"]
